core/internal/testutils: add Context; move WaitTimeout
