+++ core/internal/testutils/testutils.go
@@ -1,8 +1,11 @@
 package testutils
 
 import (
+	"context"
 	"math/big"
 	mrand "math/rand"
+	"testing"
+	"time"
 
 	"github.com/ethereum/go-ethereum/common"
 )
@@ -30,4 +33,29 @@ func randomBytes(n int) []byte {
 func Random32Byte() (b [32]byte) {
 	copy(b[:], randomBytes(32))
 	return b
+}
+
+// DefaultWaitTimeout is the default wait timeout. If you have a *testing.T, use WaitTimeout instead.
+const DefaultWaitTimeout = 30 * time.Second
+
+// WaitTimeout returns a timeout based on the test's Deadline, if available.
+// Especially important to use in parallel tests, as their individual execution
+// can get paused for arbitrary amounts of time.
+func WaitTimeout(t *testing.T) time.Duration {
+	if d, ok := t.Deadline(); ok {
+		// 10% buffer for cleanup and scheduling delay
+		return time.Until(d) * 9 / 10
+	}
+	return DefaultWaitTimeout
+}
+
+// Context returns a context with the test's deadline, if available.
+func Context(t *testing.T) (ctx context.Context) {
+	ctx = context.Background()
+	if d, ok := t.Deadline(); ok {
+		var cancel func()
+		ctx, cancel = context.WithDeadline(ctx, d)
+		t.Cleanup(cancel)
+	}
+	return ctx
 }
